add IPLog to iplog table

+++ core/src/main/java/com/gxa/p2p/common/mapper/LogininfoMapper.java
@@ -33,6 +33,6 @@ public interface LogininfoMapper {
      */
     Logininfo login(@Param("username") String username,
                     @Param("password") String password,
-                    @Param("usertype") int usertype);
+                    @Param("usertype") Byte usertype);
 
 }
