+++ core/src/main/java/com/gxa/p2p/common/service/ILoginInfoService.java
@@ -16,7 +16,7 @@ public interface ILoginInfoService {
      *  @param username
      * @param password
      */
-    Logininfo login(String username, String password, HttpServletRequest request, int usertype);
+    Logininfo login(String username, String password, HttpServletRequest request, Byte usertype);
 
 
 }
